+++ examens/views.py
@@ -5,16 +5,25 @@ from rest_framework import status
 from accounts.models  import ExamenRadiologique , ExamenBiologique , ResultatExamen
 from .serializers import ExamenRadiologiqueSerializer , ExamenBiologiqueSerializer , ResultatExamenSerializer
 from datetime import datetime
+from rest_framework.permissions import IsAuthenticated
+from rest_framework.authentication import SessionAuthentication, BasicAuthentication
+from accounts.mixin import CheckUserRoleMixin
 
-
-class ResultatExamenView(APIView):
+class ResultatExamenView(APIView,CheckUserRoleMixin):
+    permission_classes = [IsAuthenticated]
 
     def get(self, request):
+        if not self.check_user_role(request.user, ['technicien','patient'],['laborantin','medecin']):
+            return Response({'error': 'You do not have permission to create this resource.'}, status=status.HTTP_403_FORBIDDEN)
+
         resultats = ResultatExamen.objects.all()
         serializer = ResultatExamenSerializer(resultats, many=True)
         return Response(serializer.data)
 
     def post(self, request):
+        if not self.check_user_role(request.user, ['technicien'],['laborantin']):
+            return Response({'error': 'You do not have permission to create this resource.'}, status=status.HTTP_403_FORBIDDEN)
+
         serializer = ResultatExamenSerializer(data=request.data)
         if serializer.is_valid():
             serializer.save()
@@ -22,6 +31,9 @@ class ResultatExamenView(APIView):
         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
     def put(self, request, pk):
+        if not self.check_user_role(request.user, ['technicien'],['laborantin','medecin']):
+            return Response({'error': 'You do not have permission to create this resource.'}, status=status.HTTP_403_FORBIDDEN)
+
         try:
             resultat = ResultatExamen.objects.get(pk=pk)
         except ResultatExamen.DoesNotExist:
@@ -34,6 +46,9 @@ class ResultatExamenView(APIView):
         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
     def delete(self, request, pk):
+        if not self.check_user_role(request.user, ['technicien'],['laborantin']):
+            return Response({'error': 'You do not have permission to create this resource.'}, status=status.HTTP_403_FORBIDDEN)
+
         try:
             resultat = ResultatExamen.objects.get(pk=pk)
             resultat.delete()
@@ -41,14 +56,20 @@ class ResultatExamenView(APIView):
         except ResultatExamen.DoesNotExist:
             return Response({'error': 'Résultat d\'examen non trouvé'}, status=status.HTTP_404_NOT_FOUND)
 
-class ExamenBiologiqueView(APIView):
-
+class ExamenBiologiqueView(APIView,CheckUserRoleMixin):
+    permission_classes = [IsAuthenticated]
     def get(self, request):
+        if not self.check_user_role(request.user, ['technicien','patient'],['laborantin','medecin']):
+            return Response({'error': 'You do not have permission to create this resource.'}, status=status.HTTP_403_FORBIDDEN)
+
         examens = ExamenBiologique.objects.all()
         serializer = ExamenBiologiqueSerializer(examens, many=True)
         return Response(serializer.data)
 
     def post(self, request):
+        if not self.check_user_role(request.user, ['technicien'],['medecin']):
+            return Response({'error': 'You do not have permission to create this resource.'}, status=status.HTTP_403_FORBIDDEN)
+
         serializer = ExamenBiologiqueSerializer(data=request.data)
         if serializer.is_valid():
             serializer.save()
@@ -56,6 +77,9 @@ class ExamenBiologiqueView(APIView):
         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
     def put(self, request, pk):
+        if not self.check_user_role(request.user, ['technicien'],['laborantin','medecin']):
+            return Response({'error': 'You do not have permission to create this resource.'}, status=status.HTTP_403_FORBIDDEN)
+
         try:
             examen = ExamenBiologique.objects.get(pk=pk)
         except ExamenBiologique.DoesNotExist:
@@ -68,6 +92,9 @@ class ExamenBiologiqueView(APIView):
         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
     def delete(self, request, pk):
+        if not self.check_user_role(request.user, ['technicien'],['medecin']):
+            return Response({'error': 'You do not have permission to create this resource.'}, status=status.HTTP_403_FORBIDDEN)
+
         try:
             examen = ExamenBiologique.objects.get(pk=pk)
             examen.delete()
@@ -75,14 +102,22 @@ class ExamenBiologiqueView(APIView):
         except ExamenBiologique.DoesNotExist:
             return Response({'error': 'Examen Biologique non trouvé'}, status=status.HTTP_404_NOT_FOUND)
 
-class ExamenRadiologiqueView(APIView):
+class ExamenRadiologiqueView(APIView,CheckUserRoleMixin):
+    permission_classes = [IsAuthenticated]
 
     def get(self, request):
+        if not self.check_user_role(request.user, ['technicien','patient'],['radiologue','medecin']):
+            return Response({'error': 'You do not have permission to create this resource.'}, status=status.HTTP_403_FORBIDDEN)
+
         examens = ExamenRadiologique.objects.all()
         serializer = ExamenRadiologiqueSerializer(examens, many=True)
         return Response(serializer.data)
 
     def post(self, request):
+        
+        if not self.check_user_role(request.user, ['technicien'],['medecin']):
+            return Response({'error': 'You do not have permission to create this resource.'}, status=status.HTTP_403_FORBIDDEN)
+
         serializer = ExamenRadiologiqueSerializer(data=request.data)
         if serializer.is_valid():
             serializer.save()
@@ -90,6 +125,9 @@ class ExamenRadiologiqueView(APIView):
         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
     def put(self, request, pk):
+        if not self.check_user_role(request.user, ['technicien'],['radiologue','medecin']):
+            return Response({'error': 'You do not have permission to create this resource.'}, status=status.HTTP_403_FORBIDDEN)
+
         try:
             examen = ExamenRadiologique.objects.get(pk=pk)
         except ExamenRadiologique.DoesNotExist:
@@ -102,6 +140,9 @@ class ExamenRadiologiqueView(APIView):
         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
     def delete(self, request, pk):
+        if not self.check_user_role(request.user, ['technicien'],['medecin']):
+            return Response({'error': 'You do not have permission to create this resource.'}, status=status.HTTP_403_FORBIDDEN)
+
         try:
             examen = ExamenRadiologique.objects.get(pk=pk)
             examen.delete()
@@ -109,8 +150,13 @@ class ExamenRadiologiqueView(APIView):
         except ExamenRadiologique.DoesNotExist:
             return Response({'error': 'Examen Radiologique non trouvé'}, status=status.HTTP_404_NOT_FOUND)
 
-class SearchExamenBiologiqueView(APIView):
+class SearchExamenBiologiqueView(APIView,CheckUserRoleMixin):
+    permission_classes = [IsAuthenticated]
+
     def get(self, request, *args, **kwargs):
+        if not self.check_user_role(request.user, ['technicien','patient'],['laborantin','medecin']):
+            return Response({'error': 'You do not have permission to create this resource.'}, status=status.HTTP_403_FORBIDDEN)
+
         technicien = request.GET.get('technicien', None)
         date = request.GET.get('date', None)
         dossier = request.GET.get('dossier', None)
@@ -132,8 +178,13 @@ class SearchExamenBiologiqueView(APIView):
         except Exception as e:
             return Response({"detail": str(e)}, status=status.HTTP_400_BAD_REQUEST)
 
-class SearchExamenRadiologiqueView(APIView):
+class SearchExamenRadiologiqueView(APIView,CheckUserRoleMixin):
+    permission_classes = [IsAuthenticated]
+
     def get(self, request, *args, **kwargs):
+        if not self.check_user_role(request.user, ['technicien','patient'],['radiologue','medecin']):
+            return Response({'error': 'You do not have permission to create this resource.'}, status=status.HTTP_403_FORBIDDEN)
+
         technicien = request.GET.get('technicien', None)
         date = request.GET.get('date', None)
         dossier = request.GET.get('dossier', None)
@@ -156,8 +207,13 @@ class SearchExamenRadiologiqueView(APIView):
             return Response({"detail": str(e)}, status=status.HTTP_400_BAD_REQUEST)
 
 
-class SearchResultatBiologiqueByIdView(APIView):
+class SearchResultatBiologiqueByIdView(APIView,CheckUserRoleMixin):
+    permission_classes = [IsAuthenticated]
+
     def get(self, request, *args, **kwargs):
+        if not self.check_user_role(request.user, ['technicien','patient'],['laborantin','medecin']):
+            return Response({'error': 'You do not have permission to create this resource.'}, status=status.HTTP_403_FORBIDDEN)
+
         id_examen_bio = request.GET.get('idExamenBio', None)
 
         if not id_examen_bio:
@@ -177,8 +233,11 @@ class SearchResultatBiologiqueByIdView(APIView):
 
 
 class GraphiquePatientView(APIView):
-
+    permission_classes = [IsAuthenticated]
     def get(self, request, patient_id):
+        if not self.check_user_role(request.user, ['technicien','patient'],['laborantin','medecin']):
+            return Response({'error': 'You do not have permission to create this resource.'}, status=status.HTTP_403_FORBIDDEN)
+
         examens = ExamenBiologique.objects.filter(dossier_patient_id=patient_id)
 
         if not examens:
